Problem: command line argument parsing
There are a couple of libraries (or subsets of libraries) that take care
of command line argument parsing.

However, they all seem to share a common flaw:

They all use a runtime indirection. You supply a list of options and
other configuration details and run what amounts to a "blackbox"
function that will interpret your input and produce the output.

This takes away the control flow and composability. That's not great.
You can't customize it. You can't optimize it. You just put things in
and get something else out.

(I mean, most of the time this doesn't really matter as this stuff is
fairly standardized and little deviation exists. But wouldn't it be
great to make things better even if they don't matter that much?)

Solution: develop a small toolkit to handle command line parsing
in standard C, with no blackboxes.

This leaves the entire flow in the hands of the programmer while
providing small but nevertheless material assistance to the process by
implementing handling of commonly known patterns.

The library is indeed tiny and the only reason why it's over 800 LOC at
the moment is that I decided to cover it with ACSL specification to
verify the algorithms (however tiny they are) using Frama-C.

As a byproduct, code using this library generates a fairly small and
nice assembly code (subjectively speaking) that does very little. This is
not very important (we're not crunching numbers!) but I guess I am
somewhat happy about this nevertheless.