Problem: using code to satisfy WP
There is a problematic piece of code in `clam_match_signed_integer10`.
It checks if the string is empty by calling `clam_match_end`. This was
done because WP was not able to prove post-conditions on `clam_match_signed_integer10`
otherwise.

However, the annoying part about this is that the next call this
function makes is `clam_match_anychar` which calls `clam_match_end` as
its first step!

Solution: establish better post-conditions on `clam_match_anychar`
and `clam_match_end` to facilitate WP proving process.

This removes the need for `clam_match_end` call in
`clam_match_signed_integer10`.