Problem: checking counters against UINTPTR_MAX
There are a few algorithms in CLAM that have to check whether the
counter is not overflowing, just to appease Frama-C. What's even worse,
this is subtly incorrect as it then uses this counter to be added to
pointers, which may overflow at any point (at least in theory).

Solution: axiomatize string length upper boundary

This provides an assurance that string length is never greater than
UINTPTR_MAX. This is naively self-evident at the moment, although a
proof would have been nicer.