Problem: clam_match_chars_to_end specification failure
This seems to be a regression with the upgrade to Frama-C 22.0 (but I
don't have an immediate access to previous versions so can't yet prove
this)

Solution: use `strcmp` instead of `strncmp` for ensurances
(this pacifies failure of `match` behavior)
and ensure that clam_match_chars will never return anything longer than
`input` (this pacifies a failure around the requirements for a
call to `clam_match_end`).

I am not yet sure why using `strncmp` was failing as opposed to
`strcmp`.